requirements.txt
Flask
face-recognition
opencv-python
numpy
Pillow